Moved insert nested page node function to separate lib module
